requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
pillow>=9.0.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.64.0

# Data processing
pyarrow>=10.0.0
fsspec>=2023.1.0
cloudpickle>=2.0.0

# Machine Learning & AI
openai>=1.0.0
huggingface-hub>=0.15.0
pydantic>=2.0.0
pydantic_core>=2.0.0

# Web frameworks and async
httpx>=0.24.0
anyio==4.8.0
asgiref>=3.6.0

# Image processing and computer vision
opencv-python>=4.5.0

# GUI automation (Windows specific)
PyAutoGUI>=0.9.50
PyGetWindow>=0.0.9
pyperclip>=1.8.0
pytweening>=1.0.0

# System utilities
psutil>=5.8.0
docker>=6.0.0
paramiko>=3.0.0

# Cryptography and security
cryptography>=41.0.0
bcrypt>=4.0.0
PyJWT>=2.6.0

# OpenTelemetry (optional - for monitoring)
# opentelemetry-api>=1.20.0
# opentelemetry-sdk>=1.20.0
# opentelemetry-instrumentation>=0.41b0

# Development and utilities
colorama>=0.4.0
tabulate>=0.9.0
jsonschema>=4.0.0
jsonpickle>=3.0.0
strictyaml>=1.7.0

# Compatibility
typing_extensions>=4.5.0
packaging>=23.0
python-dateutil>=2.8.0
six>=1.16.0
attrs>=23.0.0

# Additional pip dependencies
annotated-types==0.7.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
idna==3.10
jeepney==0.8.0
jiter==0.8.2
mouseinfo==0.1.3
pyasn1==0.6.1
pymsgbox==1.0.9
pyopenssl==24.3.0
pyrect==0.2.0
pyscreeze==1.0.1
python3-xlib==0.15
secretstorage==3.3.3
sniffio==1.3.1

pytz==2024.1
oauthlib==3.2.2
urllib3==2.3.0
certifi==2025.1.31
bottleneck==1.4.2
numexpr==2.10.1
wheel==0.45.1
xz==5.6.4
zlib==1.2.13
yaml==0.2.5
readline==8.2
